Create Assignment6-BankAccount-Create-Data-Button
This one has the "Create Data" Button working correctly. That is what is called for in Assignment 6 for CIS 62 at Shasta College in CIS 62 Java Programming. The file above in the repository for Assignment6 needs a change in the addRecord() method.  THIS IS THE CORRECTED VERSION WHERE THE "Create Data" Button WORKS CORRECTLY.